source-custom-rss-feed/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-custom-rss-feed"
description = "Source implementation for custom RSS feed connector."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/custom-rss-feed"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_custom_rss_feed" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.9,<3.14"
airbyte-cdk = "^0"
feedparser = "*"
python-dateutil = "*"

[tool.poetry.scripts]
source-custom-rss-feed = "source_custom_rss_feed.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"

